+++ config/entrypoint.sh
@@ -3,6 +3,7 @@ set -e
 
 # Setup environment
 source $HOME/.bashrc
+source /usr/share/gazebo/setup.bash
 
 # Change owner of file
 chown $USER /home/.bash_history